PASSWORD-MANAGER-mongodb/requirements.txt
customtkinter==5.2.1
pymongo==4.6.1
python-dotenv==1.0.0
bcrypt==4.1.2
Pillow==10.2.0
tabulate==0.9.0